refactor config locations
