--- multicamera_acquisition/config/__init__.py
@@ -1,6 +1,0 @@
-from .default_basler_config import default_basler_config
-from .default_azure_config import default_azure_config
-from .default_display_config import default_display_config
-from .default_nvc_writer_config import default_nvc_writer_config
-
-# from .config import *  # causes circular import
